refactor: migrate useForm to new helper convention
